+++ clang/lib/Driver/ToolChains/Linux.cpp
@@ -1005,7 +1005,8 @@ Linux::getSupportedSanitizers(BoundArch BA,
       IsRISCV64 || IsSystemZ || IsHexagon || IsLoongArch64)
     Res |= SanitizerKind::Leak;
   if (IsX86_64 || IsMIPS64 || IsAArch64 || IsPowerPC64 || IsSystemZ ||
-      IsLoongArch64 || IsRISCV64)
+      IsLoongArch64 || IsRISCV64 ||
+      getTriple().getArch() == llvm::Triple::sparcv9)
     Res |= SanitizerKind::Thread;
   if (IsX86_64 || IsAArch64 || IsSystemZ || IsHexagon)
     Res |= SanitizerKind::Type;