+++ compiler-rt/test/tsan/map32bit.cpp
@@ -8,12 +8,7 @@
 // https://github.com/google/sanitizers/issues/412
 
 // MAP_32BIT flag for mmap is supported only for x86_64.
-// XFAIL: target=mips{{.*}}
-// XFAIL: target=aarch64{{.*}}
-// XFAIL: target=powerpc64{{.*}}
-// XFAIL: target=s390x{{.*}}
-// XFAIL: target=loongarch64{{.*}}
-// XFAIL: target=riscv64{{.*}}
+// REQUIRES: x86_64-target-arch
 
 // MAP_32BIT doesn't exist on OS X and NetBSD.
 // UNSUPPORTED: darwin,target={{.*netbsd.*}}